Migrate FdmlReader tests to munit
